remove "using strict";
The issue is when in strict mode setting `element.style = "string";` does not
work in Chrome or Safari. Both generate an error about accessing read only
value or trying to set something with only a getter. Removing the "using strict"
fixes it.

I have no idea if that's a bug in Firefox not being strict enough or a bug in
Chrome and Safari being too strict. [Asked on SO](http://stackoverflow.com/questions/32468352/uncaught-typeerror-cannot-set-property-style-of-htmlelement-which-has-only-a).